move and improve the compaction test
